+++ mailet/api/src/main/java/org/apache/mailet/Mailet.java
@@ -42,10 +42,7 @@ import com.google.common.collect.ImmutableList;
  * </ol>
  * <p>
  * In addition to the life-cycle methods, this interface provides the
- * {@link #getMailetConfig} method, which provides the Mailet with
- * its initialization parameters and a {@link MailetContext} through which
- * it can interact with the mailet container, and the {@link #getMailetInfo}
- * method, which provides basic information about the Mailet.
+ * {@link #getMailetInfo} method, which provides basic information about the Mailet.
  * <p>
  * Mailets are grouped by the mailet container's configuration into processors.
  * Each processor is comprised of an ordered sequence of Mailets, each with a
@@ -114,18 +111,12 @@ public interface Mailet {
     }
 
     /**
-     * Returns a MailetConfig object, which provides initialization parameters
-     * and a {@link MailetContext} through which it can interact with the
-     * mailet container.
-     * <p>
-     * Implementations of this interface are responsible for storing the
-     * MailetConfig which they receive in the {@link #init} method so
-     * that this method can return it.
-     *
-     * @return the MailetConfig that this mailet was initialized with
+     * @return mailet name
      */
-    MailetConfig getMailetConfig();
-    
+    default String getName() {
+        return this.getClass().getSimpleName();
+    }
+
     /**
      * Returns information about the mailet, such as author, version and
      * copyright.