+++ mailet/mailetdocs-maven-plugin/src/test/java/org/apache/james/mailet/excluded/ExcludedFromDocumentationMailet.java
@@ -42,11 +42,6 @@ public class ExcludedFromDocumentationMailet implements Mailet {
     }
 
     @Override
-    public MailetConfig getMailetConfig() {
-        return null;
-    }
-
-    @Override
     public String getMailetInfo() {
         return null;
     }