+++ mailet/mailetdocs-maven-plugin/src/test/java/org/apache/james/mailet/excluded/NotExcludedFromDocumentationMailet.java
@@ -40,11 +40,6 @@ public class NotExcludedFromDocumentationMailet implements Mailet {
     }
 
     @Override
-    public MailetConfig getMailetConfig() {
-        return null;
-    }
-
-    @Override
     public String getMailetInfo() {
         return null;
     }