+++ server/mailet/mailetcontainer-impl/src/main/java/org/apache/james/mailetcontainer/impl/ProcessorImpl.java
@@ -23,6 +23,7 @@ import java.util.List;
 import java.util.Locale;
 
 import org.apache.james.mailetcontainer.lib.AbstractStateMailetProcessor.MailetProcessorListener;
+import org.apache.james.mailetcontainer.lib.MailProcessingErrorHandlingConfiguration;
 import org.apache.james.metrics.api.MetricFactory;
 import org.apache.james.metrics.api.TimeMetric;
 import org.apache.james.util.MDCBuilder;
@@ -30,7 +31,6 @@ import org.apache.mailet.Attribute;
 import org.apache.mailet.AttributeValue;
 import org.apache.mailet.Mail;
 import org.apache.mailet.Mailet;
-import org.apache.mailet.MailetConfig;
 import org.apache.mailet.base.MailetPipelineLogging;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -44,12 +44,19 @@ public class ProcessorImpl {
     private static final Logger LOGGER = LoggerFactory.getLogger(ProcessorImpl.class);
 
     private final MetricFactory metricFactory;
+    private final MailProcessingErrorHandlingConfiguration processingErrorHandlingConfiguration;
+    private final List<MailetProcessorListener> listeners;
     private final Mailet mailet;
-    private final MailetProcessorImpl processor;
 
-    public ProcessorImpl(MetricFactory metricFactory, MailetProcessorImpl processor, Mailet mailet) {
+    public ProcessorImpl(
+            MetricFactory metricFactory,
+            Mailet mailet,
+            MailProcessingErrorHandlingConfiguration processingErrorHandlingConfiguration,
+            List<MailetProcessorListener> listeners
+    ) {
         this.metricFactory = metricFactory;
-        this.processor = processor;
+        this.processingErrorHandlingConfiguration = processingErrorHandlingConfiguration;
+        this.listeners = listeners;
         this.mailet = mailet;
     }
 
@@ -81,17 +88,8 @@ public class ProcessorImpl {
             mailet.service(mail);
         } catch (Throwable me) {
             ex = me;
-            String onMailetException = null;
+            String onMailetException = processingErrorHandlingConfiguration.onMailetException().map(value -> value.trim().toLowerCase(Locale.US)).orElse(Mail.ERROR);
 
-            MailetConfig mailetConfig = mailet.getMailetConfig();
-            if (mailetConfig instanceof MailetConfigImpl) {
-                onMailetException = mailetConfig.getInitParameter("onMailetException");
-            }
-            if (onMailetException == null) {
-                onMailetException = Mail.ERROR;
-            } else {
-                onMailetException = onMailetException.trim().toLowerCase(Locale.US);
-            }
             if (onMailetException.equalsIgnoreCase("ignore")) {
                 // ignore the exception and continue
                 // this option should not be used if the mail object can be
@@ -101,13 +99,12 @@ public class ProcessorImpl {
             } else if (onMailetException.equalsIgnoreCase("propagate")) {
                 throw me;
             } else {
-                ProcessorUtil.handleException(me, mail, mailet.getMailetConfig().getMailetName(), onMailetException, LOGGER);
+                ProcessorUtil.handleException(me, mail, mailet.getName(), onMailetException, LOGGER);
             }
 
         } finally {
             timeMetric.stopAndPublish();
             MailetPipelineLogging.logEndOfMailetProcess(mailet, mail);
-            List<MailetProcessorListener> listeners = processor.getListeners();
             long complete = System.currentTimeMillis() - start;
             if (mail.getRecipients().isEmpty()) {
                 mail.setState(Mail.GHOST);