+++ server/mailet/mailets/src/test/java/org/apache/james/samples/mailets/InstrumentationMailet.java
@@ -53,13 +53,13 @@ public class InstrumentationMailet implements Mailet {
     }
 
     @Override
-    public String getMailetInfo() {
-        return "Example mailet";
+    public String getName() {
+        return config.getMailetName();
     }
 
     @Override
-    public MailetConfig getMailetConfig() {
-        return config;
+    public String getMailetInfo() {
+        return "Example mailet";
     }
 
     @Override